requirements.txt
Django==2.1.5
wheel==0.33.1
mysqlclient
djangorestframework==3.9.2
